Refactor and setup testing
