+++ jestSetup.js
@@ -1,0 +1,7 @@
+const nodeCrypto = require('crypto');
+
+window.crypto = {
+  getRandomValues: function (buffer) {
+    return nodeCrypto.randomFillSync(buffer);
+  },
+};
