+++ package.json
@@ -56,6 +56,7 @@
     "prosemirror-tables": "^1.1.1",
     "prosemirror-utils": "^1.0.0-0",
     "scroll-into-view-if-needed": "^2.2.28",
+    "ts-jest": "^28.0.5",
     "uuid": "^8.3.2"
   },
   "peerDependencies": {
@@ -91,9 +92,11 @@
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/pickers": "^3.3.10",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^12.0.0",
     "@types/classnames": "^2.3.1",
     "@types/codemirror": "5.60.5",
-    "@types/jest": "^26.0.24",
+    "@types/jest": "^28.1.2",
     "@types/katex": "^0.11.1",
     "@types/lodash.debounce": "^4.0.6",
     "@types/lodash.isequal": "^4.5.5",
@@ -143,6 +146,8 @@
     "fuse.js": "^6.4.6",
     "html-loader": "^2.1.2",
     "html-webpack-plugin": "^5.5.0",
+    "jest": "^28.1.1",
+    "jest-environment-jsdom": "^28.1.1",
     "prettier": "^2.6.1",
     "prosemirror-collab": "^1.2.2",
     "prosemirror-model": "^1.15.0",
@@ -181,5 +186,8 @@
   },
   "files": [
     "dist"
-  ]
+  ],
+  "resolution": {
+    "react": "17.0.2"
+  }
 }
