+++ tsconfig.test.json
@@ -1,0 +1,7 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"allowJs": true
+	},
+	"include": ["test/**/*", "src/**/*", "src/global.d.ts", "demo/**/*"]
+}
